+++ Changelog.md
@@ -1,5 +1,11 @@
 ## 0.0.2
++ Added `FunctionUnity.ToAudioClip` extension method for converting a Function `Value` to an `AudioClip`.
 + Fixed Function access key not being available in builds.
++ Refactored `Feature` class to `Value` for improved clarity.
++ Refactored `PredictionService.ToFeature` method to `PredictionService.ToValue`.
++ Refactored `PredictionService.ToValue` method to `PredictionService.ToObject`.
++ Refactored `FunctionUnity.ToFeature` extension method to `FunctionUnity.ToValue`.
++ Refactored `UploadType.Feature` enumeration member to `UploadType.Value`.
 
 ## 0.0.1
 + First pre-release.